Fixing issue with BasicAuth
Passing invalid credentials to login worked as a form of "logging out" in Chrome but not Firefox.

Firefox latched onto the invalid credentials (likely because they were hardcoded in the URL) and refused
to switch to the newly entered credentials. This made things awkward.

To fix this, the invalid credentials system was replaced with a "last logged-out" session variable. This
will be set to the last user to have attempted a logout -- and cleared after the next attempted login.

This allows the first login attempt to determine that it should always fail (thereby prompting for username
and password).

Since the last user to logout is the next user trying to log in -- and we know that the user wanted to log out --
we invalidate their credentials by returning a 401 here (and prompt for the next user/pass). Crazy but it works.